% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_paths.R
\name{build_analysis_path}
\alias{build_analysis_path}
\title{Build path to analysis file or directory}
\usage{
build_analysis_path(year = NULL, filename = NULL, check_exists = FALSE)
}
\description{
Returns path to src/year/analysis/\link{filename}.
If filename is NULL, returns the directory path.
}
