% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_paths.R
\name{build_code_analysis_path}
\alias{build_code_analysis_path}
\title{Path to analysis file/dir: src/\if{html}{\out{<year>}}/analysis/\if{html}{\out{<file>}}}
\usage{
build_code_analysis_path(year = NULL, filename = NULL, check_exists = FALSE)
}
\description{
Path to analysis file/dir: src/\if{html}{\out{<year>}}/analysis/\if{html}{\out{<file>}}
}
