% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_paths.R
\name{build_code_loaders_path}
\alias{build_code_loaders_path}
\title{Path to loaders file/dir: src/\if{html}{\out{<year>}}/loaders/\if{html}{\out{<file>}}}
\usage{
build_code_loaders_path(year = NULL, filename = NULL, check_exists = FALSE)
}
\description{
Path to loaders file/dir: src/\if{html}{\out{<year>}}/loaders/\if{html}{\out{<file>}}
}
