% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_paths.R
\name{build_code_processing_path}
\alias{build_code_processing_path}
\title{Path to a processing file/dir: src/\if{html}{\out{<year>}}/processing/\if{html}{\out{<module>}}/\if{html}{\out{<state>}}/\if{html}{\out{<file>}}}
\usage{
build_code_processing_path(
  year = NULL,
  module = NULL,
  state = NULL,
  filename = NULL,
  check_exists = FALSE
)
}
\description{
Path to a processing file/dir: src/\if{html}{\out{<year>}}/processing/\if{html}{\out{<module>}}/\if{html}{\out{<state>}}/\if{html}{\out{<file>}}
}
