% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_paths.R
\name{build_data_final_path}
\alias{build_data_final_path}
\title{Path to FINAL outputs: data/\if{html}{\out{<year>}}/final/\if{html}{\out{<state>}}/\if{html}{\out{<file>}}}
\usage{
build_data_final_path(
  year = NULL,
  state = NULL,
  filename = NULL,
  check_exists = FALSE
)
}
\description{
Path to FINAL outputs: data/\if{html}{\out{<year>}}/final/\if{html}{\out{<state>}}/\if{html}{\out{<file>}}
}
