% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_paths.R
\name{build_data_processed_path}
\alias{build_data_processed_path}
\title{Path to PROCESSED data: data/\if{html}{\out{<year>}}/processed/\if{html}{\out{<module>}}/\if{html}{\out{<state>}}/\if{html}{\out{<file>}}}
\usage{
build_data_processed_path(
  year = NULL,
  module = NULL,
  state = NULL,
  filename = NULL,
  check_exists = FALSE
)
}
\description{
Path to PROCESSED data: data/\if{html}{\out{<year>}}/processed/\if{html}{\out{<module>}}/\if{html}{\out{<state>}}/\if{html}{\out{<file>}}
}
