% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_paths.R
\name{build_final_data_path}
\alias{build_final_data_path}
\title{Build path to final output data file or directory}
\usage{
build_final_data_path(
  year = NULL,
  state = NULL,
  filename = NULL,
  check_exists = FALSE
)
}
\description{
Returns path to data/year/final/\link{state}/\link{filename}.
If filename is NULL, returns the directory path.
}
