% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_paths.R
\name{build_loaders_path}
\alias{build_loaders_path}
\title{Build path to loaders file or directory}
\usage{
build_loaders_path(year = NULL, filename = NULL, check_exists = FALSE)
}
\description{
Returns path to src/year/loaders/\link{filename}.
If filename is NULL, returns the directory path.
}
