% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_paths.R
\name{build_processed_data_path}
\alias{build_processed_data_path}
\title{Build path to processed data file or directory}
\usage{
build_processed_data_path(
  year = NULL,
  module = NULL,
  state = NULL,
  filename = NULL,
  check_exists = FALSE
)
}
\description{
Returns path to data/year/processed/module/\link{state}/\link{filename}.
If filename is NULL, returns the directory path.
}
