% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_paths.R
\name{build_processing_path}
\alias{build_processing_path}
\title{Build path to a processing file or directory}
\usage{
build_processing_path(
  year = NULL,
  module = NULL,
  state = NULL,
  filename = NULL,
  check_exists = FALSE
)
}
\description{
Returns path to src/year/processing/module/\link{state}/\link{filename}.
If filename is NULL, returns the directory path.
}
