% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_paths.R
\name{build_reference_data_path}
\alias{build_reference_data_path}
\title{Path to REFERENCE data: data/\if{html}{\out{<year>}}/reference/\if{html}{\out{<module>}}/\if{html}{\out{<state>}}/\if{html}{\out{<file>}}}
\usage{
build_reference_data_path(
  year = NULL,
  module = NULL,
  state = NULL,
  filename = NULL,
  check_exists = FALSE
)
}
\description{
Path to REFERENCE data: data/\if{html}{\out{<year>}}/reference/\if{html}{\out{<module>}}/\if{html}{\out{<state>}}/\if{html}{\out{<file>}}
}
