% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_paths.R
\name{get_all_sss_dirs}
\alias{get_all_sss_dirs}
\title{Get all standard SSS directories (validates existence)}
\usage{
get_all_sss_dirs(child_care_inputs_state = "inputs")
}
\description{
Get all standard SSS directories (validates existence)
}
