% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_path.R
\name{get_base_path}
\alias{get_base_path}
\title{Get Base Path}
\usage{
get_base_path()
}
\value{
A string representing the base path for the current operating system.
}
\description{
Determines the base path based on the operating system.
Validates that the base path exists.
}
