% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{sss_check_environment}
\alias{sss_check_environment}
\title{Quick environment check for students}
\usage{
sss_check_environment()
}
\value{
(invisibly) TRUE if OK; otherwise errors with guidance.
}
\description{
Quick environment check for students
}
