% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{sss_code_base}
\alias{sss_code_base}
\title{Resolve the SSS \emph{code} base (parent folder containing your local repos)}
\usage{
sss_code_base(base = NULL, must_exist = TRUE)
}
\arguments{
\item{base}{Optional override of the base path.}

\item{must_exist}{If \code{TRUE}, error if the path doesn't exist.}
}
\value{
Absolute, normalized path (or \code{""} if unset and allowed).
}
\description{
Resolution order:
\enumerate{
\item explicit \code{base}
\item \code{getOption('sss.code')}
\item \code{Sys.getenv('SSS_CODE_BASE')}
}
}
\details{
It's OK if this is unset \emph{when} \code{repo = NULL} is used with \code{sss_code_path()}
(auto-detect current repo).
}
