% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_paths.R
\name{build_reference_data_path}
\alias{build_reference_data_path}
\title{Build path to reference data file or directory}
\usage{
build_reference_data_path(
  year = NULL,
  module = NULL,
  state = NULL,
  filename = NULL,
  check_exists = FALSE
)
}
\description{
Returns path to data/year/reference/module/\link{state}/\link{filename}.
If filename is NULL, returns the directory path.
}
